Css cleanup

+++ install.php
@@ -16,6 +16,7 @@ if (utilPhpVersion() < '5.1.2') {
 
 utilSetMagicQuotesOff();
 $error = false;
+$resetPassword = false;
 define('DEFAULT_PLUGIN', 'page');
 
 $pluginsManager = new pluginsManager();
@@ -28,6 +29,36 @@ if (isset($_GET['updateto'])) {
 			$append['defaultPlugin'] = 'page';
 			if(!saveConfig($coreConf, $append)) $error = true;
 			break;
+		case '1.2.4':
+			if (is_dir(ROOT.'data/') && file_exists(ROOT.'data/config.txt')) {
+				if(!file_exists(ROOT.'data/.htaccess')){
+					if (!@file_put_contents(ROOT.'data/.htaccess', "deny from all", 0666)) {
+						$error = true;
+					}
+				}
+				if(!file_exists(ROOT.'data/upload/.htaccess')){
+					if (!@file_put_contents(ROOT.'data/upload/.htaccess', "allow from all", 0666)) {
+						$error = true;
+					}
+				}
+				$key = uniqid(true);
+				if(!file_exists(ROOT.'data/key.php') && !@file_put_contents(ROOT.'data/key.php', "<?php define('KEY', '$key'); ?>", 0666)){
+					$error = true;
+				}
+				include(ROOT.'data/key.php');
+				$mdp = rand(100000, 999999);
+				$config = utilReadJsonFile(ROOT.'data/config.txt');
+				$config['adminPwd'] = encrypt($mdp);
+				$config['urlRewriting'] = '0';
+				utilWriteJsonFile(ROOT.'data/config.txt', $config);
+				
+				$resetPassword = true;
+			} else {
+				$error = true;
+				header('location:install.php');
+				die();
+			}
+			break;
 	}
 	
 	if ($error) {
@@ -37,40 +68,67 @@ if (isset($_GET['updateto'])) {
 		$data['msg'] = "Mise à jour effectuée";
 		$data['msgType'] = "success";
 	}
+	
+	if ($resetPassword) {
+		$data['msg'] .= "\nLe mot de passe admin a été réinitialisé : " . $mdp;
+	}
 } else {
 	if (file_exists(ROOT.'data/config.txt')) {
 		die();
 	}
 	
-	@unlink(ROOT.'.htaccess');
+	//@unlink(ROOT.'.htaccess');
+	@chmod(ROOT.'.htaccess', 0666);
 	$mdp = rand(100000, 999999); //Mot de Passe aléatoire
-	if (!@file_put_contents(ROOT.'.htaccess', "Options -Indexes", 0666)) {
+	if(!file_exists(ROOT.'.htaccess')){
+		if (!@file_put_contents(ROOT.'.htaccess', "Options -Indexes", 0666)) {
+			$error = true;
+		}
+	}
+
+	if (!is_dir(ROOT.'data/') && (!@mkdir(ROOT.'data/') || !@chmod(ROOT.'data/', 0777))) {
+		$error = true;
+	}
+	
+	if(!file_exists(ROOT.'data/.htaccess')){
+		if (!@file_put_contents(ROOT.'data/.htaccess', "deny from all", 0666)) {
+			$error = true;
+		}
+	}
+
+	if (!is_dir(ROOT.'data/plugin/') && (!@mkdir(ROOT.'data/plugin/') || !@chmod(ROOT.'data/plugin/', 0777))) {
 		$error = true;
 	}
 
-	if (!@mkdir('data/') || !@chmod('data/', 0777)) {
+	if (!is_dir(ROOT.'data/upload/') && (!@mkdir(ROOT.'data/upload/') || !@chmod(ROOT.'data/upload/', 0777))) {
 		$error = true;
 	}
-
-	if (!@mkdir('data/plugin/') || !@chmod('data/plugin/', 0777)) {
+	
+	if(!file_exists(ROOT.'data/upload/.htaccess')){
+		if (!@file_put_contents(ROOT.'data/upload/.htaccess', "allow from all", 0666)) {
+			$error = true;
+		}
+	}
+	
+	$key = uniqid(true);
+	if(!file_exists(ROOT.'data/key.php') && !@file_put_contents(ROOT.'data/key.php', "<?php define('KEY', '$key'); ?>", 0666)){
 		$error = true;
 	}
-
-	if (!@mkdir('data/upload/') || !@chmod('data/upload/', 0777)) {
-		$error = true;
-	}
+	include(ROOT.'data/key.php');
 
 	$config = array(
 		'siteName' => "Démo",
 		'siteDescription' => "Un site propulsé par 99Ko",
-		'adminPwd' => sha1($mdp),
+		//'adminPwd' => sha1($mdp),
+		'adminPwd' => encrypt($mdp), 
 		'theme' => 'defaulthtml5',
 		'adminEmail'=> 'you@domain.com',
 		'siteUrl' => getSiteUrl(),
 		'defaultPlugin' => 'page',
+		'urlRewriting' => '0',
 	);
 	
-	if(		!@file_put_contents(ROOT.'data/config.txt', json_encode($config))
+	if(!@file_put_contents(ROOT.'data/config.txt', json_encode($config))
 		||	!@chmod('data/config.txt', 0666)) {
 		$error = true;
 	}
@@ -91,7 +149,7 @@ if (isset($_GET['updateto'])) {
 		$data['msg'] = "99ko est installé\nLe mot de passe admin par défaut est : $mdp\nModifiez-le dès votre première connexion\nSupprimez également le fichier install.php";
 		$data['msgType'] = "success";
 		// On supprime le fichier d'installation et on redirige sur la page d'accueil.
-		unlink('install.php');
+		//unlink('install.php');
 	}
 }
 ?>
@@ -105,10 +163,6 @@ if (isset($_GET['updateto'])) {
 <head>
        <meta charset="utf-8">  
        <title>99ko - Installation</title>
-       <!-- meta -->
-       <meta name="description" content="Cms hyper léger!">
-       <meta name="author" content="Jonathan C.">
-       <meta name="generator" content="99Ko">
        <!-- css -->
        <link rel="stylesheet" href="admin/css/style.css" media="all">
        <link rel="stylesheet" href="admin/css/common.css" media="all">
@@ -131,18 +185,18 @@ if (isset($_GET['updateto'])) {
 </head>
 <body>
  
-       <aside>
+       <!--<aside>
                <div id="copyright">
                   Propulsé par <a target="_blank" title="CMS sans base de données" href="http://99ko.tuxfamily.org/">99ko</a> <span class="version"><?php echo $data['99koVersion']; ?></span>.
                </div>
-       </aside>
+       </aside>-->
        
        <div id="content">      
 <section id="home">
-       <h1>Installation</h1>
+       <!--<h1>Installation</h1>
        <h2><a class="btn" id="logout" href="./admin/">Administration</a>
        <a class="btn" id="showSite" href="./">Voir le site</a></h2>
-       <hr>
+       <hr>-->
        <?php showMsg($data['msg'], $data['msgType']); ?>
 </section>
     </div>
